+++ src/main/java/frc/team4909/robot/subsystems/climber/ClimberSubsystem.java
@@ -10,8 +10,8 @@ import edu.wpi.first.wpilibj.drive.DifferentialDrive;
 import frc.team4909.robot.RobotMap;
 import frc.team4909.robot.RobotConstants;
 import frc.team4909.robot.subsystems.drivetrain.commands.Drive;
-import frc.team4909.robot.subsystems.climber.ExtendStilts;
-import frc.team4909.robot.subsystems.climber.RetractStilts;
+import frc.team4909.robot.subsystems.climber.commands.ExtendStilts;
+import frc.team4909.robot.subsystems.climber.commands.RetractStilts;
 import com.ctre.phoenix.motorcontrol.FeedbackDevice;
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import frc.team4909.robot.Robot;
