+++ src/main/java/frc/team4909/robot/subsystems/climber/commands/ExtendStilts.java
@@ -1,16 +1,16 @@
-package frc.team4909.robot.subsystems.climber;
+package frc.team4909.robot.subsystems.climber.commands;
 
 import edu.wpi.first.wpilibj.command.InstantCommand;
 import frc.team4909.robot.Robot;
 import frc.team4909.robot.RobotConstants;
 
-public class RetractStilts extends InstantCommand {
-    public RetractStilts() {
+public class ExtendStilts extends InstantCommand {
+    public ExtendStilts() {
         requires(Robot.climberSubsystem);
     }
 
     protected void execute() {
-        Robot.climberSubsystem.retractStilts();
+        Robot.climberSubsystem.extendStilts();
     }
 
 }
