+++ src/main/java/frc/team4909/robot/subsystems/climber/commands/StopExtend.java
@@ -1,4 +1,4 @@
-package frc.team4909.robot.subsystems.climber;
+package frc.team4909.robot.subsystems.climber.commands;
 
 import edu.wpi.first.wpilibj.command.InstantCommand;
 import frc.team4909.robot.Robot;
